riskcast-v6/requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.15.0
kaleido==0.2.1
openpyxl>=3.0.10
fpdf2>=2.7.0
Pillow>=9.0.0
statsmodels>=0.13.0
